Made GameState, improved liquid, integrated lighting, made layers, removed SWING graphics support
There is a bug in the liquids where making  a lot of particles freezes the game.  I suspect it may have
something to do with the fountain, since if the number of particles spawned is more than one per
update loop, the game also freezes.  Now updated Components can access a GameState instance,
reducing the need for passing references.  Layers are now up and running.  Lighting looks great, but I
still have to test normals.